working chaotic ant, but could be better

+++ include/edge.h
@@ -4,7 +4,7 @@
 using namespace std;
 
 typedef struct edge{
-	float val; 
+	int val; 
 	int neighbour;	
 	edge() : val(0), neighbour(0){}
 	edge(float new_val, int new_neighbour) : val(new_val), neighbour(new_neighbour){}
